+++ baseball_talent_manager/frontend/src/views/manager/DashboardView.vue
@@ -26,10 +26,9 @@
           </div>
         </div>
       </div>
-
     </div>
   </div>
 </template>
 
-<script setup>
-</script>
+<!-- <script setup>
+</script> -->